launch runtime compiles
