+++ polkadot-parachains/launch-runtime/src/weights/pallet_utility.rs
@@ -16,8 +16,7 @@
 // --execution=wasm
 // --wasm-execution=compiled
 // --heap-pages=4096
-// --output=polkadot-parachains/launch-runtime/src/weights/pallet_utility.rs
-
+// --output=polkadot-parachains/encointer-runtime/src/weights/pallet_utility.rs
 
 #![cfg_attr(rustfmt, rustfmt_skip)]
 #![allow(unused_parens)]
@@ -44,5 +43,10 @@ impl<T: frame_system::Config> pallet_utility::WeightInfo for WeightInfo<T> {
 	}
 	fn dispatch_as() -> Weight {
 		(31_300_000 as Weight)
+	}
+
+	fn force_batch(_c: u32) -> Weight {
+		// Todo: dummy weight need to rerun benchmarks
+		(23_000_000 as Weight)
 	}
 }
